requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-timeout==3.0.1
attrs==21.2.0
Brotli==1.0.9
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
Deprecated==1.2.13
discord.py==1.7.3
ffmpeg-python==0.2.0
frozenlist==1.2.0
future==0.18.2
idna==3.3
multidict==5.2.0
mutagen==1.45.1
packaging==21.3
pycryptodomex==3.15.0
pyparsing==3.0.9
python-dotenv==0.19.2
redis==3.5.3
requests==2.26.0
six==1.16.0
spotipy==2.20.0
typing-extensions==4.0.1
urllib3==1.26.7
websockets==10.3
wrapt==1.14.1
yarl==1.7.2
youtube-search==2.1.1
yt-dlp==2022.8.14
